Lich su ban nick, fix bug mua nick

+++ app/Http/Controllers/AccountController.php
@@ -48,7 +48,10 @@ class AccountController extends Controller
     public function show($categorySlug, $accountUuid)
     {
         $category = Category::where('slug', $categorySlug)->first();
-        $account = Account::with('images')->whereIn('status', [Account::STATUS_SOLD, Account::STATUS_AVAILABLE])->where('uuid', $accountUuid)->first();
+        $account = Account::with('images')
+            ->whereIn('status', [Account::STATUS_SOLD, Account::STATUS_AVAILABLE])
+            ->where('uuid', $accountUuid)
+            ->first();
 
         if (empty($category) || empty($category) || !empty($category) && $category->status !== Category::ACTIVE_STATUS || empty($account)) {
             return redirect()->route('home');
@@ -173,7 +176,7 @@ class AccountController extends Controller
             ->where('uuid', $uuid)
             ->first();
 
-        if (!empty($account)) {
+        if (!empty($account) && $account->status !== Account::STATUS_AVAILABLE) {
             $account->delete();
             return redirect()->back()->with('success', 'Xóa thành công!');
         }
